Uneeded files removed
